Refactor ORM helpers
