+++ src/lib/orm/dialect/index.ts
@@ -1,21 +1,25 @@
 import type { TableRelations } from "../orm/types.js";
 import type { Table } from "../table/types.js";
-import { createPostgresDialect } from "./postgres.js";
-import { createSqliteDialect } from "./sqlite.js";
+import { SqlDialect } from "./dialect.js";
+import { POSTGRES_SPEC } from "./postgres.js";
+import { SQLITE_SPEC } from "./sqlite.js";
 import type { Adapter, Dialect, GetDialectInput } from "./types.js";
 
 export const getDialect = <T extends Table, R extends TableRelations>(
   input: GetDialectInput<T, R>,
 ): Dialect<T, R> => {
-  const { adapter, table, relations, tableRelationsMap = new Map() } = input;
+  const dialectInput = {
+    ...input,
+    tableRelationsMap: input.tableRelationsMap ?? new Map(),
+  };
 
-  switch (adapter) {
+  switch (input.adapter) {
     case "sqlite":
-      return createSqliteDialect({ table, relations, tableRelationsMap });
+      return new SqlDialect({ spec: SQLITE_SPEC, ...dialectInput });
     case "postgres":
-      return createPostgresDialect({ table, relations, tableRelationsMap });
+      return new SqlDialect({ spec: POSTGRES_SPEC, ...dialectInput });
     default:
-      throw new Error(`Unsupported adapter: ${adapter}`);
+      throw new Error(`Unsupported adapter: ${input.adapter}`);
   }
 };
