+++ src/lib/orm/dialect/postgres.ts
@@ -1,7 +1,4 @@
-import type { TableRelations } from "../orm/types.js";
-import type { Table } from "../table/types.js";
-import { createDialect } from "./shared.js";
-import type { CreateNamedDialectInput, DialectSpec } from "./types.js";
+import type { DialectSpec } from "./types.js";
 
 export const POSTGRES_SPEC: DialectSpec = {
   name: "postgres",
@@ -11,10 +8,3 @@ export const POSTGRES_SPEC: DialectSpec = {
   jsonArrayAggregateFunctionName: "jsonb_agg",
   emptyJsonArrayLiteral: "'[]'::jsonb",
 };
-
-export const createPostgresDialect = <
-  T extends Table,
-  R extends TableRelations,
->(
-  input: CreateNamedDialectInput<T, R>,
-) => createDialect({ spec: POSTGRES_SPEC, ...input });
