+++ src/lib/orm/dialect/sqlite.ts
@@ -1,7 +1,4 @@
-import type { TableRelations } from "../orm/types.js";
-import type { Table } from "../table/types.js";
-import { createDialect } from "./shared.js";
-import type { CreateNamedDialectInput, DialectSpec } from "./types.js";
+import type { DialectSpec } from "./types.js";
 
 export const SQLITE_SPEC: DialectSpec = {
   name: "sqlite",
@@ -11,7 +8,3 @@ export const SQLITE_SPEC: DialectSpec = {
   jsonArrayAggregateFunctionName: "json_group_array",
   emptyJsonArrayLiteral: "'[]'",
 };
-
-export const createSqliteDialect = <T extends Table, R extends TableRelations>(
-  input: CreateNamedDialectInput<T, R>,
-) => createDialect({ spec: SQLITE_SPEC, ...input });
